apps/admin: Rename write_ins.ts files -> adjudication.ts, add getOptionCoordinates helper, refactor getting ballot image to remove write-in coupling
