+++ apps/admin/frontend/src/screens/tally/manual_tallies_form_screen.tsx
@@ -60,7 +60,7 @@ import {
   getWriteInCandidates,
   setManualResults,
 } from '../../api';
-import { normalizeWriteInName } from '../../utils/write_ins';
+import { normalizeWriteInName } from '../../utils/adjudication';
 import {
   ManualTallyFormContestParams,
   ManualTallyFormParams,
