--- apps/admin/frontend/src/utils/write_ins.ts
@@ -1,3 +1,0 @@
-export function normalizeWriteInName(name: string): string {
-  return name.toLowerCase().trim().replace(/\s+/, ' ');
-}
